Use OpenShift's wildcard certs for the secure-sso route
Take advantage of the OpenShift router's valid certificate and only
use the self-signed certificate for internal communication from the
OpenShift router to the SSO server. This gives us valid certificates
externally, assuming the OpenShift router is setup with a valid
wildcard certificate. And because we're reencrypting we get TLS
security all the way through to the SSO server.